Removed utils.py added fliplr and rotate to accufiz config, used newer submit_data function.
