+++ catkit2/services/accufiz_interferometer_sim/accufiz_interferometer_sim.py
@@ -9,22 +9,26 @@ import math
 from astropy.io import fits
 from glob import glob
 from catkit2.testbed.service import Service
-from catkit2.utils import rotate_and_flip_image
 import tempfile
 import os
 import threading
-
 
 class sim_response:
     text = 'success'
 
 def generate_circle_array(radius=1, h=256, w=256):
-    # radius is the proportion of the circle (0.5 is half the size)
-    # h is the height of the rectangle
-    # w is the width of the rectangle array with circle
-    # the purpose of this function is to generate a dummy mask
+    """Generate a dummy mask as numpy ndarray of size (h, w)
 
-    # Create a grid of points with shape (grid_size, grid_size)
+    Args:
+        radius (int, optional): radius is the proportion of the circle (0.5 is half the size). Defaults to 1.
+        h (int, optional): the height of the rectangle. Defaults to 256.
+        w (int, optional): the width of the rectangle array with circle. Defaults to 256.
+
+    Returns:
+        numpy.ndarray uint8: dummmy mask
+    """
+
+    # Create a grid of points with shape (h, w)
     x = np.linspace(-1, 1, w)
     y = np.linspace(-1, 1, h)
     xx, yy = np.meshgrid(x, y)
@@ -37,9 +41,25 @@ def generate_circle_array(radius=1, h=256, w=256):
 
     return circle_array.astype('uint8')
 
+def rotate_and_flip_image(data, theta, flip):
+        """
+        Converts an image based on rotation and flip parameters.
+        :param data: Numpy array of image data.
+        :param theta: Rotation in degrees of the mounted camera, only these discrete values {0, 90, 180, 270}
+        :param flip: Boolean for whether to flip the data using np.fliplr.
+        :return: Converted numpy array.
+        """
+        data_corr = np.rot90(data, int(theta / 90))
+
+        if flip:
+            data_corr = np.fliplr(data_corr)
+
+        return data_corr
 
 class AccufizInterferometerSim(Service):
+
     NUM_FRAMES_IN_BUFFER = 20
+    instrument_lib = requests
 
     def __init__(self):
         super().__init__('accufiz_interferometer_sim')
@@ -52,7 +72,6 @@ class AccufizInterferometerSim(Service):
         # these are the optional configurations and will automatically default to something for convenience
         self.ip = self.config.get('ip_address', 'localhost:8080')
         self.calibration_data_package = self.config.get('calibration_data_package', '')
-        self.sim_data = self.config.get('sim_data', None)
         self.timeout = self.config.get('timeout', 10000)
         self.post_save_sleep = self.config.get('post_save_sleep', 1)
         self.file_mode = self.config.get('file_mode', True)
@@ -62,6 +81,9 @@ class AccufizInterferometerSim(Service):
         self.save_h5 = self.config.get('save_h5', True)
         self.save_fits = self.config.get('save_fits', False)
         self.num_frames_avg = self.config.get('num_avg', 2)
+        self.fliplr = self.config.get('fliplr', True)
+        self.rotate = self.config.get('rotate', 0)
+        self.sim_data = self.config.get('sim_data', None)
 
         # Set the 4D timeout.
         self.html_prefix = f"http://{self.ip}/WebService4D/WebService4D.asmx"
@@ -84,8 +106,6 @@ class AccufizInterferometerSim(Service):
         self.make_command('start_acquisition', self.start_acquisition)
         self.make_command('end_acquisition', self.end_acquisition)
 
-    instrument_lib = requests
-
     def set_mask(self):
         # Set the Mask. This mask has to be local to the 4D computer in this directory.
         filemask = self.mask
@@ -97,16 +117,12 @@ class AccufizInterferometerSim(Service):
 
         return True
 
-
-    # TODO: could store a realistic list of POST/GET response expectations. But need to store the commands as retreived from the hardware.
     def get(self, url, params=None, **kwargs):
         return sim_response()
-
 
     def post(self, url, data=None, json=None, **kwargs):
         time.sleep(self.post_save_sleep)
         return sim_response()
-
 
     def take_measurement(self):
         # Send request to take data.
@@ -132,9 +148,9 @@ class AccufizInterferometerSim(Service):
         else:
             local_file_path = self.sim_data.replace('.h5', '')
 
-        #  This line is here because when sent through webservice slashes tend
-        #  to disappear. If we sent in parameter a path with only one slash,
-        #  they disappear
+        # This line is here because when sent through webservice slashes tend
+        # to disappear. If we sent in parameter a path with only one slash,
+        # they disappear
         server_file_path = server_file_path.replace('\\', '/')
         server_file_path = server_file_path.replace('/', '\\\\')
 
@@ -153,7 +169,7 @@ class AccufizInterferometerSim(Service):
 
         self.detector_masks.submit_data(mask.astype(np.uint8))
 
-        image = self.convert_h5_to_fits(local_file_path, rotate=0, fliplr=True, mask=mask, img=img, create_fits=self.save_fits)
+        image = self.convert_h5_to_fits(local_file_path, rotate=self.rotate, fliplr=self.fliplr, mask=mask, img=img, create_fits=self.save_fits)
         if temp_file:
             fitsfile = local_file_path.replace('.h5', '.fits')
 
@@ -209,21 +225,16 @@ class AccufizInterferometerSim(Service):
                 has_correct_parameters = np.allclose(self.images.shape, img.shape)
 
                 if not has_correct_parameters:
-                    self.images.update_parameters('uint16', img.shape, 20)
-                self.log.info('requesting new frame')
-                frame = self.images.request_new_frame()
+                    self.images.update_parameters('float32', img.shape, 20)
 
-                frame.data[:] = img
-                self.log.info(f'frame {frame.id} acquired')
-                self.images.submit_frame(frame.id)
+                self.images.submit_data(img.astype('float32'))
+
                 time.sleep(0.05)
         finally:
             self.is_acquiring.submit_data(np.array([0], dtype='int8'))
 
-
     def start_acquisition(self):
         self.should_be_acquiring.set()
-
 
     def end_acquisition(self):
         self.should_be_acquiring.clear()
