--- catkit2/utils.py
@@ -1,16 +1,0 @@
-import numpy as np
-
-def rotate_and_flip_image(data, theta, flip):
-    """
-    Converts an image based on rotation and flip parameters.
-    :param data: Numpy array of image data.
-    :param theta: Rotation in degrees of the mounted camera, only these discrete values {0, 90, 180, 270}
-    :param flip: Boolean for whether to flip the data using np.fliplr.
-    :return: Converted numpy array.
-    """
-    data_corr = np.rot90(data, int(theta / 90))
-
-    if flip:
-        data_corr = np.fliplr(data_corr)
-
-    return data_corr
